+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 2022-11-08
 
+### Added
+
+- Overscan settings to scale entire viewport down
+
 ### Changed
 
 - Restructured the package to work with Unity's Package Manager